BK KEY: lindgren@debian.as|ChangeSet|20050106170254|62597

+++ src/include/xmms/ipc_msg.h
@@ -28,7 +28,7 @@ typedef struct xmms_ipc_msg_St {
 	guint32 object;
 	guint32 cmd;
 	guint32 cid;
-	guint16 get_pos, data_length;
+	guint32 get_pos, data_length;
 	guint8 *data;
 	guint32 size;
 } xmms_ipc_msg_t;
